pyproject.toml
[tool.poetry]
name = "news-bot"
version = "0.1.0"
description = "A bot that fetches and summarizes news from various sources"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
